backend/requirements.txt
Django==5.0.*
celery[redis]==5.3.6
celery-redbeat==2.1.1
django-model-utils==4.3.1
django-webpack-loader==3.1.0
django-js-reverse==0.10.2
django-import-export==3.3.5
djangorestframework==3.14.0
django-debreach==2.1.0
python-decouple==3.8
python-dotenv==1.0.1
psycopg2==2.9.9
brotlipy==0.7.0
django-log-request-id==2.1.0
dj-database-url==2.1.0
whitenoise==6.6.0
psutil==5.9.7
ipython==8.18.1
sentry-sdk==1.39.1
setuptools==69.0.2
django-permissions-policy==4.18.0
django-csp==3.7
django-defender==0.9.7
ply==3.11
django-guid==3.5.0
drf-spectacular==0.27.2
gunicorn==22.0.0
django-unfold==0.24.0
django-anymail[mailgun]==10.3
djangorestframework-simplejwt[crypto]==5.3.1